requirements.txt
asgiref==3.8.1
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2025.4
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.2.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.2.1
django-appconf==1.1.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-filter==25.1
django-htmx==1.23.0
django-humanize==0.1.2
django-import-export==4.3.7
django-plotly-dash==2.4.6
django-select2==8.4.0
django-tables2==2.7.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dpd-static-support==0.0.5
dpd_components==0.1.0
et_xmlfile==2.0.0
executing==2.2.0
Faker==37.1.0
Flask==3.0.3
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gunicorn==21.2.0
httplib2==0.22.0
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==1.38.1
nest-asyncio==1.6.0
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
plotly==6.0.1
pluggy==1.5.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.30.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
reportlab==4.4.0
requests==2.32.3
retrying==1.3.4
rsa==4.9.1
setuptools==80.3.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tablib==3.8.0
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.0.6
whitenoise==6.9.0
XlsxWriter==3.2.3
zipp==3.21.0
